requirements.txt
# Core dependencies
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Retry and resilience
tenacity>=8.2.0

# Logging and monitoring
colorlog>=6.7.0

# Configuration validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI
click>=8.1.7

# Date/time handling
python-dateutil>=2.8.2

# Security (optional - for advanced credential management)
# cryptography>=41.0.0
# azure-keyvault-secrets>=4.7.0
# azure-identity>=1.15.0
